pyproject.toml
[tool.poetry]
name = "final-class"
version = "0.0.1"
description = "Final classes for Python 3"
authors = ["sobolevn <mail@sobolevn.me>"]
license = "MIT"

readme = "README.md"

repository = "https://github.com/wemake-services/final-class"
homepage = "https://github.com/wemake-services/final-class"

keywords = ["mimesis", "pytest", "testing"]

[tool.poetry.dependencies]
python = "^3.6 || ^3.7"

[tool.poetry.dev-dependencies]
pytest-isort = "^0.2.0"
pytest-flake8 = "^1.0"
pytest-randomly = "^1.2"
pytest-cov = "^2.5"
flake8-builtins = "^1.4"
flake8-commas = "^2.0"
flake8-quotes = "^1.0"
flake8-pep3101 = "^1.2"
flake8-comprehensions = "^1.4"
flake8-blind-except = "^0.1.1"
flake8-docstrings = "^1.3"
flake8-string-format = "^0.2.3"
flake8-coding = "^1.3"
flake8-module-name = "^0.1.5"
flake8-bugbear = "^18.2"
flake8-pytest = "^1.3"
pep8-naming = "^0.7.0"
flake8 = "^3.5"
mypy = "^0.610.0"
